Move Qwen component trace helpers out of main
