+++ crates/runner/src/main.rs
@@ -78,12 +78,12 @@ use qwen35_runtime::{
     Qwen35EngineSetup, Qwen35Policy, Qwen35Prefill, Qwen35Startup,
 };
 use qwen35_trace::{
-    trace_component_input_layer, trace_persistent_input_layer,
+    trace_component_input_layer, trace_component_layer, trace_component_linear_layer,
+    trace_component_linear_state_layer, trace_persistent_input_layer,
     trace_persistent_linear_state_layer,
 };
 use qwen35_trace_utils::{
-    bf16_residual_sum, bf16_round, build_linear_decode_v_reference, fp8_e4m3_to_f32_host,
-    sigmoid_fast,
+    bf16_residual_sum, bf16_round, fp8_e4m3_to_f32_host, sigmoid_fast,
 };
 use registry::{Backend, FamilyParams, GpuArch, ModelFamily, ModelVariant};
 use supersonic_core::backend::{BackendChoice, BACKEND_CHOICES};
@@ -2973,260 +2973,6 @@ fn trace_persistent_linear_layer(
 
     eprintln!(
         "[trace-persistent-linear] layer={trace_layer} hidden_delta={hidden_delta:.6} comp_vs_replay_conv={comp_vs_replay_conv:.6} comp_vs_replay_recurrent={comp_vs_replay_recurrent:.6} comp_linear_hidden_vs_replay={comp_vs_replay_hidden:.6} native_vs_comp_qkv_proj={native_vs_comp_qkv_proj:.6} native_vs_replay_qkv_proj={native_vs_replay_qkv_proj:.6} native_vs_comp_z_proj={native_vs_comp_z_proj:.6} native_vs_replay_z_proj={native_vs_replay_z_proj:.6} native_vs_comp_b_proj={native_vs_comp_b_proj:.6} native_vs_replay_b_proj={native_vs_replay_b_proj:.6} native_vs_comp_a_proj={native_vs_comp_a_proj:.6} native_vs_replay_a_proj={native_vs_replay_a_proj:.6} native_vs_comp_conv={native_vs_comp_conv:.6} native_vs_comp_recurrent={native_vs_comp_recurrent:.6} native_conv_vs_expected_tail={native_conv_vs_expected_tail:.6} comp_conv_vs_expected_tail={comp_conv_vs_expected_tail:.6} replay_conv_vs_expected_tail={replay_conv_vs_expected_tail:.6} native_conv_tap_deltas=[{native_conv_tap_deltas}] comp_conv_tap_deltas=[{comp_conv_tap_deltas}] max_append_mismatch=({max_append_mismatch}) step_b_debug=({step_b_debug}) first_later_clobber=({first_later_clobber}) pointer_debug=({pointer_debug}) isolated_tail_windows=[{isolated_tail_windows}] append_mismatch_samples=[{append_mismatch_samples}] native_vs_comp_token_mixer={native_vs_comp_token_mixer:.6} native_vs_replay_token_mixer={native_vs_replay_token_mixer:.6} native_vs_comp_post_norm={native_vs_comp_post_norm:.6} native_vs_replay_post_norm={native_vs_replay_post_norm:.6} native_vs_comp_gated={native_vs_comp_gated:.6} native_vs_replay_gated={native_vs_replay_gated:.6} native_vs_comp_swiglu={native_vs_comp_swiglu:.6} native_vs_replay_swiglu={native_vs_replay_swiglu:.6} native_vs_comp_mlp_down={native_vs_comp_mlp_down:.6} native_vs_replay_mlp_down={native_vs_replay_mlp_down:.6} native_vs_comp_proj_residual={native_vs_comp_proj_residual:.6} comp_layer_hidden_vs_replay={comp_layer_vs_replay_hidden:.6} native_vs_comp_layer_hidden={native_vs_comp_layer_hidden:.6} native_vs_replay_hidden={native_vs_replay_hidden:.6} sample_qkv_native=[{sample_qkv_native}] sample_qkv_comp=[{sample_qkv_comp}] sample_z_native=[{sample_z_native}] sample_z_comp=[{sample_z_comp}]"
-    );
-    Ok(())
-}
-
-fn trace_component_layer(
-    engine: &DecodeEngine,
-    trace_layer: usize,
-    native: &decode_engine::ComponentLayerTrace,
-    token_ids: &[u32],
-    ordinal: usize,
-    kv_chunk_size: usize,
-    prefill_chunk_size: usize,
-    use_4b_kernel: bool,
-) -> Result<()> {
-    let replay = prefill_engine::prefill(
-        engine.weights(),
-        &mut ModelState::new(&engine.weights().config, ordinal)
-            .map_err(|e| anyhow::anyhow!("component layer trace replay state init: {e}"))?,
-        engine.rotary(),
-        token_ids,
-        ordinal,
-        kv_chunk_size,
-        prefill_chunk_size,
-        false,
-        use_4b_kernel,
-        true,
-        None,
-    )?;
-    let attn = replay
-        .layer_attn_trace
-        .as_ref()
-        .and_then(|layers| layers.get(trace_layer))
-        .ok_or_else(|| anyhow::anyhow!("missing replay attn trace for layer {trace_layer}"))?;
-    let post = replay
-        .layer_post_attn_norm_trace
-        .as_ref()
-        .and_then(|layers| layers.get(trace_layer))
-        .ok_or_else(|| anyhow::anyhow!("missing replay post-attn trace for layer {trace_layer}"))?;
-    let mlp = replay
-        .layer_mlp_out_trace
-        .as_ref()
-        .and_then(|layers| layers.get(trace_layer))
-        .ok_or_else(|| anyhow::anyhow!("missing replay mlp trace for layer {trace_layer}"))?;
-    let hidden = replay
-        .layer_hidden_trace
-        .as_ref()
-        .and_then(|layers| layers.get(trace_layer))
-        .ok_or_else(|| anyhow::anyhow!("missing replay hidden trace for layer {trace_layer}"))?;
-    let attn_delta =
-        validate::max_abs_delta(&decode_bf16_le(&native.attn_hidden), &decode_bf16_le(attn));
-    let post_delta = validate::max_abs_delta(
-        &decode_bf16_le(&native.post_attn_norm),
-        &decode_bf16_le(post),
-    );
-    let mlp_delta = validate::max_abs_delta(&decode_bf16_le(&native.mlp_out), &decode_bf16_le(mlp));
-    let hidden_delta = validate::max_abs_delta(
-        &decode_bf16_le(&native.layer_hidden),
-        &decode_bf16_le(hidden),
-    );
-    eprintln!(
-        "[trace-component-layer] layer={trace_layer} attn_delta={attn_delta:.6} post_norm_delta={post_delta:.6} mlp_delta={mlp_delta:.6} hidden_delta={hidden_delta:.6}"
-    );
-    Ok(())
-}
-
-fn trace_component_linear_layer(
-    engine: &DecodeEngine,
-    trace_layer: usize,
-    native: &decode_engine::ComponentLinearTrace,
-    token_ids: &[u32],
-    ordinal: usize,
-    kv_chunk_size: usize,
-    prefill_chunk_size: usize,
-    use_4b_kernel: bool,
-) -> Result<()> {
-    let replay = prefill_engine::prefill(
-        engine.weights(),
-        &mut ModelState::new(&engine.weights().config, ordinal)
-            .map_err(|e| anyhow::anyhow!("component linear trace replay state init: {e}"))?,
-        engine.rotary(),
-        token_ids,
-        ordinal,
-        kv_chunk_size,
-        prefill_chunk_size,
-        false,
-        use_4b_kernel,
-        false,
-        Some(trace_layer),
-    )?;
-    let replay = replay
-        .linear_debug_trace
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("missing replay linear trace for layer {trace_layer}"))?;
-    let qkv_delta =
-        validate::max_abs_delta(&decode_bf16_le(&native.qkv), &decode_bf16_le(&replay.qkv));
-    let z_delta = validate::max_abs_delta(&decode_bf16_le(&native.z), &decode_bf16_le(&replay.z));
-    let packed_native = decode_f32_le(&native.packed);
-    let packed_replay = decode_f32_le(&replay.packed);
-    let packed_delta = validate::max_abs_delta(&packed_native, &packed_replay);
-    let cfg = &engine.weights().config;
-    let nv = cfg.linear_num_value_heads;
-    let khd = cfg.linear_key_head_dim;
-    let vhd = cfg.linear_value_head_dim;
-    let packed_width = 2 * khd + vhd + 2;
-    let mut q_delta = 0.0f32;
-    let mut k_delta = 0.0f32;
-    let mut v_delta = 0.0f32;
-    let mut beta_delta = 0.0f32;
-    let mut gexp_delta = 0.0f32;
-    let v_ref = build_linear_decode_v_reference(engine, trace_layer, &native.qkv)?;
-    let mut v_ref_native_delta = 0.0f32;
-    let mut v_ref_replay_delta = 0.0f32;
-    let mut state_vs_tail_delta = 0.0f32;
-    if !replay.qkv_tail.is_empty() {
-        let state = engine
-            .state_for_batch(0)
-            .layers
-            .get(trace_layer)
-            .ok_or_else(|| anyhow::anyhow!("missing state for layer {trace_layer}"))?;
-        let conv_state = decode_bf16_le(
-            &state
-                .conv_state
-                .as_ref()
-                .ok_or_else(|| anyhow::anyhow!("layer {trace_layer} missing conv_state"))?
-                .to_host_bytes()
-                .map_err(|e| anyhow::anyhow!("trace conv_state D2H: {e}"))?,
-        );
-        let qkv_tail = decode_bf16_le(&replay.qkv_tail);
-        let qkv_dim = cfg.linear_num_key_heads * cfg.linear_key_head_dim * 2
-            + cfg.linear_num_value_heads * cfg.linear_value_head_dim;
-        let state_len = cfg.linear_conv_kernel_dim - 1;
-        let mut expected = vec![0.0f32; qkv_dim * state_len];
-        for t in 0..state_len {
-            for c in 0..qkv_dim {
-                expected[c * state_len + t] = qkv_tail[t * qkv_dim + c];
-            }
-        }
-        state_vs_tail_delta = validate::max_abs_delta(&conv_state, &expected);
-    }
-    for h in 0..nv {
-        let base = h * packed_width;
-        q_delta = q_delta.max(validate::max_abs_delta(
-            &packed_native[base..base + khd],
-            &packed_replay[base..base + khd],
-        ));
-        k_delta = k_delta.max(validate::max_abs_delta(
-            &packed_native[base + khd..base + 2 * khd],
-            &packed_replay[base + khd..base + 2 * khd],
-        ));
-        v_delta = v_delta.max(validate::max_abs_delta(
-            &packed_native[base + 2 * khd..base + 2 * khd + vhd],
-            &packed_replay[base + 2 * khd..base + 2 * khd + vhd],
-        ));
-        let v_ref_base = h * vhd;
-        v_ref_native_delta = v_ref_native_delta.max(validate::max_abs_delta(
-            &packed_native[base + 2 * khd..base + 2 * khd + vhd],
-            &v_ref[v_ref_base..v_ref_base + vhd],
-        ));
-        v_ref_replay_delta = v_ref_replay_delta.max(validate::max_abs_delta(
-            &packed_replay[base + 2 * khd..base + 2 * khd + vhd],
-            &v_ref[v_ref_base..v_ref_base + vhd],
-        ));
-        beta_delta = beta_delta
-            .max((packed_native[base + 2 * khd + vhd] - packed_replay[base + 2 * khd + vhd]).abs());
-        gexp_delta = gexp_delta.max(
-            (packed_native[base + 2 * khd + vhd + 1] - packed_replay[base + 2 * khd + vhd + 1])
-                .abs(),
-        );
-    }
-    let rec_apply_delta = validate::max_abs_delta(
-        &decode_f32_le(&native.rec_apply),
-        &decode_f32_le(&replay.rec_apply),
-    );
-    let attn_delta =
-        validate::max_abs_delta(&decode_bf16_le(&native.attn), &decode_bf16_le(&replay.attn));
-    let gated_delta = validate::max_abs_delta(
-        &decode_bf16_le(&native.gated),
-        &decode_bf16_le(&replay.gated),
-    );
-    let proj_out_delta = validate::max_abs_delta(
-        &decode_bf16_le(&native.proj_out),
-        &decode_bf16_le(&replay.proj_out),
-    );
-    eprintln!(
-        "[trace-component-linear] layer={trace_layer} qkv_delta={qkv_delta:.6} z_delta={z_delta:.6} packed_delta={packed_delta:.6} q_delta={q_delta:.6} k_delta={k_delta:.6} v_delta={v_delta:.6} state_vs_tail_delta={state_vs_tail_delta:.6} v_ref_native_delta={v_ref_native_delta:.6} v_ref_replay_delta={v_ref_replay_delta:.6} beta_delta={beta_delta:.6} gexp_delta={gexp_delta:.6} rec_apply_delta={rec_apply_delta:.6} attn_delta={attn_delta:.6} gated_delta={gated_delta:.6} proj_out_delta={proj_out_delta:.6}"
-    );
-    Ok(())
-}
-
-fn trace_component_linear_state_layer(
-    engine: &DecodeEngine,
-    trace_layer: usize,
-    history_token_ids: &[u32],
-    ordinal: usize,
-    kv_chunk_size: usize,
-    prefill_chunk_size: usize,
-    use_4b_kernel: bool,
-) -> Result<()> {
-    let native_layer = engine
-        .state_for_batch(0)
-        .layers
-        .get(trace_layer)
-        .ok_or_else(|| anyhow::anyhow!("missing native layer {trace_layer}"))?;
-    let native_conv = native_layer
-        .conv_state
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("layer {trace_layer} has no conv_state"))?
-        .to_host_bytes()
-        .map_err(|e| anyhow::anyhow!("native conv_state D2H: {e}"))?;
-    let native_rec = native_layer
-        .recurrent_state
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("layer {trace_layer} has no recurrent_state"))?
-        .to_host_bytes()
-        .map_err(|e| anyhow::anyhow!("native recurrent_state D2H: {e}"))?;
-
-    let mut replay_state = ModelState::new(&engine.weights().config, ordinal)
-        .map_err(|e| anyhow::anyhow!("component linear state replay init: {e}"))?;
-    prefill_engine::prefill(
-        engine.weights(),
-        &mut replay_state,
-        engine.rotary(),
-        history_token_ids,
-        ordinal,
-        kv_chunk_size,
-        prefill_chunk_size,
-        false,
-        use_4b_kernel,
-        false,
-        None,
-    )?;
-    let replay_layer = replay_state
-        .layers
-        .get(trace_layer)
-        .ok_or_else(|| anyhow::anyhow!("missing replay layer {trace_layer}"))?;
-    let replay_conv = replay_layer
-        .conv_state
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("replay layer {trace_layer} has no conv_state"))?
-        .to_host_bytes()
-        .map_err(|e| anyhow::anyhow!("replay conv_state D2H: {e}"))?;
-    let replay_rec = replay_layer
-        .recurrent_state
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("replay layer {trace_layer} has no recurrent_state"))?
-        .to_host_bytes()
-        .map_err(|e| anyhow::anyhow!("replay recurrent_state D2H: {e}"))?;
-
-    let conv_delta =
-        validate::max_abs_delta(&decode_bf16_le(&native_conv), &decode_bf16_le(&replay_conv));
-    let rec_delta =
-        validate::max_abs_delta(&decode_f32_le(&native_rec), &decode_f32_le(&replay_rec));
-    eprintln!(
-        "[trace-component-linear-state] layer={trace_layer} conv_delta={conv_delta:.6} recurrent_delta={rec_delta:.6}"
     );
     Ok(())
 }
